+++ structure/HeavyLightDecomposition.cpp
@@ -2,33 +2,33 @@
 #include "../template/template.cpp"
 
 struct HLD{
-	using V=vector<pair<int,P>>;
+	using V=vector<pair<ll,P>>;
 	struct heavy_set{
-		vector<int>ele;
-		int depth,par,cost1=1,cost2=1;
-		heavy_set(int v,int d,int par)
+		vector<ll>ele;
+		ll depth,par,cost1=1,cost2=1;
+		heavy_set(ll v,ll d,ll par)
 		:ele(1,v),depth(d),par(par){}
 	};
-	vector<vector<int>>G;
+	vector<vector<ll>>G;
 	vector<heavy_set>S;
-	vector<int>subsize,stidx,eleidx;
-	int subtree(int v,int p){
-		int &sz=subsize[v];
+	vector<ll>subsize,stidx,eleidx;
+	ll subtree(ll v,ll p){
+		ll &sz=subsize[v];
 		if(sz)return sz;
 		sz=1;
-		for(int i:G[v])if(i!=p)sz+=subtree(i,v);
+		for(ll i:G[v])if(i!=p)sz+=subtree(i,v);
 		return sz;
 	}
-	void make_path(int v,int p,int id){
+	void make_path(ll v,ll p,ll id){
 		stidx[v]=id;
 		eleidx[v]=S[id].ele.size()-1;
-		int mxidx,mx=0;
-		for(int i:G[v])if(i!=p){
+		ll mxidx,mx=0;
+		for(ll i:G[v])if(i!=p){
 			if(mx<subtree(i,v)){
 				mx=subtree(i,v);mxidx=i;
 			}
 		}
-		for(int i:G[v])if(i!=p){
+		for(ll i:G[v])if(i!=p){
 			if(mxidx==i){
 				S[id].ele.push_back(i);
 				make_path(i,v,id);
@@ -38,10 +38,10 @@ struct HLD{
 			}
 		}
 	}
-	int st(int v){return stidx[v];}
-	int el(int v){return eleidx[v];}
-	HLD(vector<vector<int>>&G):G(G){
-		int N=G.size();
+	ll st(ll v){return stidx[v];}
+	ll el(ll v){return eleidx[v];}
+	HLD(vector<vector<ll>>&G):G(G){
+		ll N=G.size();
 		subsize.resize(N);
 		eleidx.resize(N);
 		stidx.resize(N);
